refactor connect verison logic
